requirements.txt
# ======================================================
# 🚀 Core
# ======================================================
streamlit>=1.20
pandas>=1.5
numpy>=1.25

# ======================================================
# 📊 Visualización
# ======================================================
plotly>=5.0
matplotlib>=3.5

# ======================================================
# 📈 Excel y manejo de archivos
# ======================================================
xlsxwriter>=3.1
openpyxl>=3.0

# ======================================================
# 🔒 Seguridad / Auth
# ======================================================
bcrypt>=4.0

# ======================================================
# 🗄 Base de datos
# ======================================================
sqlalchemy>=2.0
psycopg2-binary>=2.9

# ======================================================
# 🔄 Manejo de archivos y concurrencia
# ======================================================
filelock>=3.12

# ======================================================
# 🖨 PDF / Reportes
# ======================================================
reportlab>=3.6

# ======================================================
# 🔧 Utilidades varias
# ======================================================

python-dateutil>=2.8
python-dotenv>=1.0